requirements.txt
attrs==20.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
idna==3.3
loguru==0.5.3
pydantic==1.9.0
PyYAML==6.0
requests==2.27.1
tqdm==4.62.3
typer==0.4.0
typing-extensions==4.1.1
urllib3==1.26.8
